Corregir errores pero aun faltan, prueba de commit

+++ src/main/java/com/duoc/stimy/gamehub/dto/TicketRequestDTO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class TicketRequestDTO {
 
     @NotNull(message = "El ID del usuario es obligatorio")
-    private Integer usuarioId;
+    private Long usuarioId; // CORREGIDO: Cambiado de Integer a Long
 
     @NotBlank(message = "El asunto es obligatorio")
     private String asunto;
